+++ Changelog.md
@@ -5,6 +5,7 @@
 Features:
 
 - File upload via drag and drop into the folder view
+- Commits that are less than a minute ago are automatically ammended
 
 Fixed: